refactor: centralize generate metadata
