+++ src/promptflow/promptflow/_utils/flow_utils.py
@@ -5,7 +5,7 @@ import hashlib
 import os
 from os import PathLike
 from pathlib import Path
-from typing import Union
+from typing import Tuple, Union
 
 from promptflow._sdk._constants import DAG_FILE_NAME, DEFAULT_ENCODING
 from promptflow._utils.logger_utils import LoggerFactory
@@ -46,16 +46,37 @@ def get_flow_lineage_id(flow_dir: Union[str, PathLike]):
     return lineage_id
 
 
-def resolve_flow_path(flow_path: Path):
-    """Resolve given flow path to dag file path."""
-    if flow_path.is_dir():
-        flow_path = flow_path / DAG_FILE_NAME
-    return flow_path
+def resolve_flow_path(
+    flow_path: Union[str, Path, PathLike], base_path: Union[str, Path, PathLike, None] = None
+) -> Tuple[Path, str]:
+    """Resolve flow path and return the flow directory path and the file name of the target yaml.
+
+    :param flow_path: The path of the flow directory or the flow yaml file. It can either point to a
+      flow directory or a flow yaml file.
+    :type flow_path: Union[str, Path, PathLike]
+    :param base_path: The base path to resolve the flow path. If not specified, the flow path will be
+      resolved based on the current working directory.
+    :type base_path: Union[str, Path, PathLike]
+    :return: The flow directory path and the file name of the target yaml.
+    :rtype: Tuple[Path, str]
+    """
+    if base_path:
+        flow_path = Path(base_path) / flow_path
+    else:
+        flow_path = Path(flow_path)
+
+    if flow_path.is_dir() and (flow_path / DAG_FILE_NAME).is_file():
+        return flow_path, DAG_FILE_NAME
+    elif flow_path.is_file():
+        return flow_path.parent, flow_path.name
+
+    raise FileNotFoundError(f"Can't find flow with path {flow_path.as_posix()}.")
 
 
 def load_flow_dag(flow_path: Path):
     """Load flow dag from given flow path."""
-    flow_path = resolve_flow_path(flow_path)
+    flow_dir, file_name = resolve_flow_path(flow_path)
+    flow_path = flow_path / file_name
     if not flow_path.exists():
         raise FileNotFoundError(f"Flow file {flow_path} not found")
     with open(flow_path, "r", encoding=DEFAULT_ENCODING) as f:
@@ -65,7 +86,7 @@ def load_flow_dag(flow_path: Path):
 
 def dump_flow_dag(flow_dag: dict, flow_path: Path):
     """Dump flow dag to given flow path."""
-    flow_path = resolve_flow_path(flow_path)
-    with open(flow_path, "w", encoding=DEFAULT_ENCODING) as f:
+    flow_dir, flow_filename = resolve_flow_path(flow_path)
+    with open(flow_dir / flow_filename, "w", encoding=DEFAULT_ENCODING) as f:
         dump_yaml(flow_dag, f)
     return flow_path
